Refactor Topbar
