+++ resources/js/components/layout/MainLayout/MainLayout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom'
-import Topbar from '@/js/components/layout/Topbar'
+import Topbar from './components/Topbar'
 
 const MainLayout = () => (
     <div className='bg-gray-50 min-h-screen text-gray-700 flex flex-col'>
