+++ resources/js/components/layout/MainLayout/components/NavLink.jsx
@@ -1,0 +1,14 @@
+import { Button, Tooltip, } from 'antd'
+import { Link } from 'react-router-dom'
+
+const NavLink = ({ to, title, icon }) => {
+    return (
+        <Link to={to}>
+            <Tooltip title={title}>
+                <Button className='rounded-lg' icon={icon} />
+            </Tooltip>
+        </Link>
+    )
+}
+
+export default NavLink
