+++ resources/js/components/layout/MainLayout/index.js
@@ -1,0 +1,1 @@
+export { default } from './MainLayout'
